recommender_service/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.23.2
requests>=2.31.0
pandas>=2.1.1
numpy>=1.26.0
python-dotenv>=1.0.0
pydantic>=2.4.2
aiohttp>=3.8.5
backoff>=2.2.1 